Add script for bootstrapping resources in database
See issue #262

Some resources need to be added by default to Krake, before it is
started. For instance, in production, roles need to be added to have a
secure infrastructure. These roles have to be manually added for Krake
to work.

A new bootstrapping/bootstrap script is added, that reads YAML files to
avoid doing this task manually. A few definitions are also added at the
same time: roles that should be added by default if RBAC is set, and
prometheus and static metrics and metric providers, for testing
purposes.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>