Ignore the reenqueuing of deleted Applications
See issue #311

When deleting an Application, the KubernetesController first updates a
few elements of the Application. It means the API is called, the
resource changes in the databases, and the KubernetesController is aware
of the update through the watch mechanism. The Application is thus
reenqueued, while it is processed for deletion. When the deletion is
finally done, the Application in deletion is released from the queue,
while the updated version is handled. However, as the resource is
deleted from the database, it cannot be handled, and the first update
during deletion raise a 404.

A quick fix has been added, to check if the Application is present in
the database or not before attempting to delete it.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>