Change the bootstrapping behavior to a rollback
See comments on MR!296

The previous bootstrap mechanism tried to add all resources, and would
not add the ones that raise an error. It means in case of failures, the
database is not in the same state before and after using the script.

With the new rollback mechanism, in case of failure, the database is put
back to the same state it had before running the script. This allows
having a cleaner database in case of failures.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>