Refactor the KubernetesObserver to use hooks
See discussions on MR !290

The Kubernetes observer were stopped before handling a resource, and a
new one was started if needed after handling. It was done with a special
care, with two functions added in the workflow only for this. Howver,
hooks are present to handle such general cases.

New hooks have thus been added for the handling of Applications, and
some have been used for registering and unregistering the
KubernetesObserver. These hooks can be reused later on with other
mechanisms, and prevent a specific behavior for the observers only.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>