Fix bootstrapping validation issue on serialization
With the new feature of having the whole data sent after serialization,
new instances need the `creation_ignored` boolean set to True during
serialization to allow readonly fields to be empty.

In the bootstrapping, this flag was not set, and the missing readonly
fields led to a ValidationError for any resources. This is now fixed.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>