Update roles for bootstrapping with Magnum resources
If the RBAC authentication mode is used along with TLS support,
different roles need to be put in the database for the different Krake
componenent.

One has been added for the Magnum Controller, and the rules of some
other roles have been updated to take into account the OpenStack
resources (Project and MagnumCluster).

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>