Fix scheduling with constraints, without metrics
In the case of resources A scheduled without metrics on resources B,
the selected resource B was taken randomly from all possible resources
B. This was the case even if the resource A had constraints. Hence, a
resource B that was not matching the constraints of a resource A could
still be selected.

Now, the resource B is selected only from the resources that matched
with the constraints.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>