Add get before update Application in Rok
See issue #322

When updating an Application with rok, an error occurred because the
readonly and subresource fields of the Application were missing. This
led to a validation error, which prevented the update to happen.

To prevent this, the Application pieces of information are first
fetched, then used for the update request.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>