Remove resources unused parts in Dependency Graph
For internal functions of the garbage collector, the dependency graph
stored for each resources the complete set of attributes of this
resources. This implied that a big part of the database was somehow
stored by the garbage collector.

This has been refactored, in order to have the dependency graph only
storing as few elements as possible from the resources that are managed.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>