Add timeout for metrics fetching
The integration test were failing in the case of two clusters, one with
a valid metrics provider and one with a metrics provider that is
defective. The process of fetching the metrics for the faulty provider
led to a long timeout (5 minutes). This led to a blocked worker and an
application not scheduled. The test thus could not see the application
in a RUNNING state fast enough, and failed.

Now, a temporary timeout has been added in the fetch process, to prevent
the workers in the scheduler to wait too much time for a metric.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>