Change the default rok user for consistency
In the roles of bootstrapping and the default configuration files, the
administrator user is called "system:admin". The default user of rok was
"system".

If rok commands are run in a directory without a rok configuration file,
the default rok user is used ("system"). When deploying, the default
configuration user is used ("system:admin").

This led for instance to the following case: by connecting to the VM
created and running rok commands, the user was "system", but the
playbooks used "sytem:admin". So the resources created by the tests
could not be seen by a user that were to launch commands in the VM
directly.

For consistency, the default rok user is changed to "system:admin".

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>