Fix updates not applied by the KubernetesController
See issue #336

When an Application is updated, it should first go through the
Scheduler, to be rescheduled if necessary. Only then the
KubernetesController should handle the update and/or the updated
scheduled cluster (and perform migration). To create this "priority"
mechansim, the "modified" and "scheduled" timestamps of an Application
are used. The former, if bigger than the other, implies that the
Application has just been created or updated on the API. In this case,
the Scheduler should handle the Application, and the
KubernetesController should reject it. During the handling on the
Scheduler, the "scheduled" timestamp is updated, and becomes bigger than
the "modified" timestamp. It is then the turn of the
KubernetesController to handle the Application. In the end, this
mechanism simply compares which of the two timestamp is bigger on each
Controller, and acts accordingly.

In the case of an update of an Application on the API, the Application
would be handled by the Scheduler, and not by the KubernetesController
thanks to this mechansim. If no better cluster was found for the
Application, the Scheduler would not update the "scheduled_to" cluster.
However, the "scheduled" timestamp was not updated either, which
prevented the KubernetesController from handling the Application and
performing updates on the Kubernest cluster.

This has now been fixed, by updating the "scheduled" timestamp on the
Scheduler, and sending an update request to the API, in the case of an
update on the Application that did not lead to an update of the
scheduled cluster.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>