Clean remaining tasks when a Controller stops
See issue #338

The background tasks of a Controller (reflectors, workers...) are all
started concurrently using the asyncio "gather" mechanism. Gather does
not cancel by itself all remaining tasks if one of them fail (the
documentation is misleading on this). It means that if one task fails,
all others still runs on the asyncio loop. When the controller stops,
all this tasks are still present and are stopped abruptly, which
displays many error messages. Mostly, the workers still try to get a
value from the queue, even if it has been closed.

This cleaning issue has now been fixed by doing this cleanup of the
remaining tasks manually.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>