Fix timestamp issue in the Metadata factory in tests
See issue #316

The creation and modified timestamps generated in the Metadata factory
for the unit tests can be generated with the same value, which makes
sense, as they have the same value when a resource is created on the
API.

However a test is performed in the factory which should prevent both to
have the same value. This can happen with a very low probability in the
tests.

Thus, the test has been modified to allow the timestamps to be equal.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>