Fix cycle detection on the DependencyGraph on update
See issue #356

In the DependencyGraph of the GarbageCollector, resources can be added,
removed or updated, which changes the relations between the current
resources and the one processed. In the case of an update, the
dependents of this resource are temporary stored, the resource is
removed from the graph, then added again with its new owners, and
finally its dependents are added again. The issue was that the cycle
detection occurred before the dependents were added again, which
sometimes implied no cycle detection.

This issue has now been fixed by putting the cycle detection only at the
very end of a resource update, and modifying the test to avoid false
positives.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>