Add API generator base and definition generation
See #303

The base code for the API generator has been added, as well as the
module for the generation of an API definition using a Krake data
module.

See #364

The generated definitions have been merged with the existing ones, thus
the unnecessary endpoint for the Kubernetes Cluster status has been
removed.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>