Store resource_version of k8s resources
The resource_version is an attribute present in all k8s resources metadata,
which tracks the version of the resource. We now store the list of resource
versions of all resources that we create. We use the resource_version in the
KubernetesObserver to detect if the resource was changed on the cluster.

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>