Remove legacy elements from infra and ansible
See issue #241

The playbooks for the deployment of a central IDP, for the magnum
clusters, for the prometheus VM and the devstack part have been removed.
References to these playbooks or these deployment parts have been
removed from the other deployments playbooks. These elements were legacy
that are not used anymore, or need to be adapted. To keep a clean
repository, they have been completely removed.

The infra/ directory has also been removed completely for the same
reason, as it is not used with the current setup.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>