Add replicas:2 to echo-demo-update in user stories
There was an inconsistency in the basics user story. As in the text the
file echo-demo-update.yaml was cat-ed, the old content of that file
was different from what was shown. In particular 'replicas: 2' was
missing, which is added. Also add integration tests checking
that the number of replicas is changed after an update.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>