Add the port as parameter of the API
The port that the API was listening to was defined only by aiohttp: 8443
if TLS was enabled, 8080 otherwise. However, in the Ansible playbooks,
we let the possibility to set the port, which was not propagated to the
API component.

This has now been fixed, and the port is now a parameter to set in the
configuration, or as command-line argument.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>