Enable Krake authorization in the deployment
See issues #250 and #366

To be able to tests the deployment of a Krake infrastructure closer to a
production-ready infrastructure, the deployment now uses RBAC as an
authorization mode. This allows to be notified if the authorization
mechanism is not working, and is a first step for end-to-end tests of
this mechanism.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>