Adapt the rok deployment to use TLS
See issues #250 and #366

For rok to be able to communicate with Krake, its configuration has been
changed to use TLS.

Also, the configuration was previously put into the git directory, which
was pointed by the XDG_CONFIG_HOME environment variable. This variable
was added in .profile to be exported. However, the change was not taken
into account as .profile was not sourced again. Thus, the configuration
of rok has been placed in "/etc/rok", another directory where rok looks
for its configuration.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>