Change the term rank to score to avoid confusion
In the scheduling, the score should correspond to the value computed
using the normalization of the sum of the metrics multiplied by the
weights. The ranks are given by the ranking step, where the scores of
all resources are compared: the ranks conceptually go from 1 to N, N
being the number of resources to schedule to the highest score gets the
highest rank, the lowest score get the lowest rank and two resources
with the same score have the same rank. The actual rank is never
computed, only the resources with highest scores get the first rank,
which is then chosen.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>