Fix observer loop with kubernetes fields on camelCase
When getting the current state of the resources it observes from the
real world, the observer tries to compare all current values with the
ones received from the actual cluster. The kubernetes_asyncio API
however converts all fields from camelCase to snake_case. Then, during
the comparison, the value of a field with camelCase is not found on the
received elements, and is considered empty.

For instance, if .apiVersion changed on the actual cluster (impossible
but let's go on), the observer would attempt to compare "apiVersion" to
the value of the "apiVersion" of the resource on the cluster. The value
of the field "apiVersion" would be searched, but does not exist in the
values given by the kubernetes API. The API converted "apiVersion" to
"api_version". As the observer does not find any value for "apiVersion",
it considered it did not change, and keeps the older value. Thus, the
observer failed to observe a change.

This has now been fixed: the observer converts the name of the fields
into snake_case before the actual comparison.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>