Fix deletion of Application without complete hook
If the "complete" hook is not enabled on an Application, the endpoint
for the hook is still present. With the previous state, the token sent
for verification could be let empty ("null" in JSON), which would be
compared to the one stored in the database, "None" if the hook is not
enabled. As the value "null" would be converted to "None", this token
would be accepted as the right one, and the Application would be
deleted.

This commit adds a first check when a request to the hook endpoint is
received. Every request sent to the "complete" hook endpoint for an
Application without the hook enabled is rejected as unauthorised.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>