Fix rescheduling of resources without a spec field
See issue #358 and #365

For the reconciliation of the Applications, the current state of the
Application (its status) is compared to its desired state (its spec).
During this process, the Kubernetes resources that exists and should be
kept are compared to the actual ones, to check if any difference is
present. However, only the spec of the Kubernetes resources were
compared. If the resource had no spec attribute, the comparison raised
an exception. This was not the expected behavior. This case arose for
instance with the "complete" hook, where a ConfigMap is added. The
ConfigMap do not have any spec attribute.

Thus, the new process is simply to compare the whole desired resource
with the current one, without focusing on the spec only.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>