Remove venv from Dockerfile and clean
See issue #375

The Dockerfile for Krake uses the python:3.6-slim image as base. We
installed inside a virtual environment using venv. This led to issues
with the installation of Krake and of the scripts (interpreter wrongly
set).

To clean the Krake image, the virtual environment has been removed,
which also solve the issue of the wrong interpreter set in the scripts.
Thus, the image is simpler, as Krake is directly installed in the image.
This removes an unnecessary layer around it.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>