Migrate only apps where migration is enabled
When an application is created it is automatically scheduled to a
cluster. Due to various reasons, the system could then decide to
reschedule the application. There was previously no way of prohibiting
this rescheduling of an application.

A user can now specify --enable-migration or --disable-migration
(rok CLI) or app:spec:constraints:migration: True/False (krake API)
when creating and/or updating a kubernetes application.
--disable-migration prevents the application from being rescheduled
or migrated. When ineither --disable-migration not --enable-migration
is specified migration is enabled.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>