Replace factories import to use tests module
The unit tests of Krake previously imported the factories for resources
creation with "from factories import ..." or "from tests.factories
import ...". The former was sometimes not recognised by linters or IDE,
which gave an error, even though the tests could be started.

For consistency, the tests now all use the second version, with the
"tests.factories". For this purpose, the "tests" directory has to be
"converted" into a module, hence the new "__init__.py" file.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>