Add external endpoint for the complete hook
See issue #365

The "complete" hook hands over the Krake API endpoint to the
Applications in its environment variables. The value of the endpoint was
given by the endpoint to which the Kubernetes controller was connected.
However, this endpoint may be private if the controller and the API are
on the same machine for instance. This endpoint is then not accessible
for an Application running on a cluster in a different machine.

To solve this, an additional parameter called "external_endpoint" can
be specified in the "complete" hook configuration in the Kubernetes
controller. If set, the value of this parameter will be given to the
Applications with the "complete" hook enabled. This will replace the
Krake API endpoint handed over, and allow Application running on a
machine not in the one where the API is running to reach the API.

To ensure the validity of the certificate from an external device, the
external endpoint is also appended to the list of valid hostnames of
the API when generating the certificates.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>