Prevent processing after non-changing rescheduling
When an Application was rescheduled, it is taken from the worker queue
of the Scheduler to be processed again. If the scheduling decision did
not change, the scheduling timestamp was updated. This triggered an
update event on the API, and the Application was received by the
KubernetesController, to be reprocessed. However, it means that, even if
the Application was not modified otherwise, the Application would go
through the reconciliation loop.

So if there are is a rescheduling, there are two cases:
 1. if the Application needs to be rescheduled, the binding is updated,
    which triggers the update of the Application, which goes to the K8s
    Controller.

 2. the scheduling decision did not change. There are again two cases:
    1. The rescheduling was triggered by an update of the Application.
       The Application still needs to be processed by the K8s
       Controller, so the scheduling timestamp is updated.
    2. The rescheduling was automatic. The Application did not change.
       The scheduling timestamp was still updated, which triggered a new
       processing of the Application by the K8s Controller.

The last point should not occur. If the Application has not been
modified, and the scheduling decision is the same, then the K8s
Controller should not process the Application.

To prevent this case, if the scheduling decision did not change, the
scheduling timestamp is only updated for an Application that was
updated, to let it be processed by the K8s Controller. If nothing was
updated, the scheduled timestamp is higher than the modified timestamp.
In this case, the scheduled timestamp is now not updated a second time,
to prevent the K8s Controller from processing it vainely.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>