Change the docker installation source in ansible
Docker-ce was previously installed, which needed some special care for
dependencies, adding a GPG key and a repository.

As the docker.io official package of the Debian/Ubuntu repository is
just a few versions behind, it is sufficient to install this version
instead of the upstream docker-ce.

See this article: https://www.collabora.com/news-and-blog/blog/2018/07/04/docker-io-debian-package-back-to-life/

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>