Add unit test migrating a kube app multiple times
It has been observed that the migrate_application() method of
the kubernetes controller consistently is called too many times
when an application is scheduled to move to a differnet cluster.
In an attempt to pin point this problem, add a unit test where
an application is created, and then migrated between clusters
multiple times.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>