Add new convenience functionality for e2e testing
When using the e2e test classes ApplicationDefinition and
ClusterDefinition, one had to first create a command
and then run it in order to update, create and/or delete the
application or cluster. Now one can use the convenience
methods update(), create() and delete() to perform those
actions directly.

Furthermore, add metrics functionality to
ClusterDefinition, since we need to be able to
create and update clusters with metrics.

We also add convenience method Environment.get_resource()
to be able to retrieve a resource by name.

Also, in the rok kube app update command one can specify
labels using the -l/--label option. This was previously
not possible using ApplicationDefinition.update_command().
Add the label option to the ApplicationDefinition.update()
method.

Add the ability to adapt the environment in which
we run the commands in the run() method.

Add a method get_scheduling_score() for calculating the
scheduling score of a cluster based on its weights,
the current metrics from the static provider and the stickiness.

Add a convenience method set_static_metrics() for changing
the metrics of the statis metrics provider.

Add a convenience method ApplicationDefinition.get_running_on()
to get the name of the cluster the application is running on.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>