Add stdin as input for the Krake bootstrap script
The `krake_bootstrap_db` script could previously only read inputs from
files given as arguments. This implied that the files should be
accessible by the script in the filesystem. However, in the case of
Krake being in a container for instance, the files to read needed then
to be added inside the container, for the script to read them.

The script can now read inputs from stdin. This allows the "-i" option
of docker exec to be used, for instance to run the bootstrap of
resource without adding their file (where they are defined) to a
container.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>