Separate docker volumes and set permissions
See issue #366

The "krake" user inside the Krake Docker image could not access the
files mounted in it, as they got the wrong owner.

Also, as we used to mount the whole "/etc/krake" into the different
containers, each container could access the certificate key of the other
components.

Thus, a specific UID for the user "krake" is now set inside the
container. This UID is then used as owner of the files mounted into the
different containers that makes the Krake components.

This also means that the certificates previously used for rok would be
owned by another user as "krake". This has now been changed, rok gets
access to the needed files (certificates, key and configuration)
actually separated in "/etc/rok".

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>