Implement observer_schema logic in k8s Observer and Controller
This commits implements the logic described in #374:
- During the reconciliation of an application the Kubernetes Controller update (or
initialize) its status.last_applied_manifest, as a copy of spec.manifest augmented by
the observed fields for which a value is known.
- During the reconciliation of a resource, the Kubernetes Controller checks if there are
difference between the values of the observed fields in the last_applied_manifest and
the last_observed_manifest.
- After the creation of update of a Kubernetes resource, the Controller updates the
status.last_applied_manifest and status.last_observed_manifest, using the Kubernetes
response as an input. The last_applied_manifest is augmented with observed fields, and
the last_observed_manifest is overwritten, so that it always holds the last known value
of the observed fields.
- The Kubernetes Observer updates the observed fields value in last_observed_manifest.

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>