Add validation of the labels of the resources
See issue #373

Previously, any kind of labels could be given to the API when creating
or updating a resource.

To restrain the possibilities and prevent edge cases with special
characters, a validation is now performed on the key and value of each
label. The accepted strings follow the same rules as for Kubernetes
labels, see
https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#syntax-and-character-set

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>