Add support of more characters in LabelConstraints
See issue #373

Previously, the `LabelConstraints` limited the label key and values,
forcing them to only consist of letters. However, in the labels
themselves, numbers and some other characters were accepted.

Thus, additional characters are added to the grammar that is used to
parse the label constraints.

The regular expression of the labels is also reused, to prevent code
duplication. It means that the keys and values parsed in the label
constraints must follow the same rules as the ones in the labels.

Because the regular expression could not be used directly in the grammar
(because of ambiguity), the validation is performed at serialization and
deserialization, and thus uses our marshmallow-based serialisation
system.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>