Refactor API to use a single database client
See issue #370

Previously, the Krake API created one etcd client for each request
received. Because of the etcd client used, a memory leakage appeared,
where the client would put too much elements in cache, see
https://github.com/Revolution1/etcd3-py/issues/141

To circumvent the issue, one single client is now used for the whole
Krake API. As this client leverages a pool of connections, it could
make sense to only use one of them. With this method, the API does not
show any sign of memory leak.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>