Allow the API to be started without database
See issue #370

The etcd client attempts a connection with the database when created.
With the single etcd client for the whole API, it is created at the
creation of the API. Thus, a database not accessible makes the API
fail at creation.

To prevent this, the etcd client is created only when the first request
is received on the API. Like this, even if the session could not be
created, the API does not stops, and the connection to the database is
attempted at the next request.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>