Add tests for migrating by cluster constraints
When the cluster label constraints of an application are
updated, a migration of the application should occur,
if the new constraints match a different cluster than the
one the application is currently running on.
The previous test only tested this behaviour using
the flags --enable-migration/--disable-migration,
but did not verify the default behaviour.

Add a test to verify this default behaviour.

Add a test to verify that the update comes at the
time of the database modification, and not at the
time of the reoccuring rescheduling interval.

We also improve the implementation of no_migration test by
removing an unnecessary wait (since check_running_on()
automatically waits and retries) and use the convenience method
app.update() instead of run(app.update_command()).

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>