Support Kubernetes non-default namespaces
See issue #396

Previously, any Kubernetes resource deployed using Krake would be
created in the `default` namespace, even in the case where a namespace
was explicitely defined in the manifest file given to Krake. This was
due to a hardcoded value for the namespace (to `default`).

To support different Kubernetes namespaces with Krake, the namespace
where each resource should be deployed is now taken from the resource
directly, and is only set to `default` when no namespace is provided.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>