Move the Kubernetes client elements to a new file
Previously, all classes which handled directly the communication with
the Kubernetes clusters were present in the
`krake.controller.kubernetes.kubernetes` module. This created some
import cycles, for instance for the `KubernetesObserver`. Thus, the
Kubernetes client had to be given as reference.

This has been modified, to have a separate file for each class that
manages the communication with the cluster. This way, the client is made
accessible for all other parts of the Kubernetes controller easily.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>