Improve output of rok
The output of most commands in rok would show the stack trace in case of
errors. Also, a user did not have the possibility to change the format
for all commands.

These two problems were fixed, and additional attributes of the
resources are also shown in the output with the "table" format.

As rok also sometimes returned an error code of 1 even though the
command was executed without problems, workarounds were put in the
integration tests Environment. They have now been removed.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>