Change command methods to return list instead of str
Previously the commands returned by creation_command and
update_command in ApplicationDefinition returned the commands as
strings. This caused a bug since commands containing option
flags like "-L location is DE" would be incorrectly executed
by the run method in utils.py.

Now the methods return a list with each part of the commands as elements:
[...., "-L", "location is DE", ...].

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>