Add support of CORS rules
When attempting to reach the Krake API through a browser, issues
regarding the CORS mechanism (Cross-origin resource sharing) could
occur, as they were not supported.

To allow the Krake API to be called by a browser, CORS rules need to be
set. They allow the API to be more publicly accessible.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>