Fix scheduler if all providers are unreachable
Previously the scheduler failed to schedule an application
to a cluster, whenever all metrics providers used by the
matching clusters were unreachable.

Instead, we now pick the matching cluster with the highest
score, in case all metrics providers are unreachable.

We also add e2e test tests this scenario and unit tests
for select_kubernetes_cluster().

The added unit tests test the above scenario and also
verifies the scheduler's behavior when _some_
metrics providers are unreachable. Unit tests are added
for both the case when the application is already
scheduled and when it is not, i.e, with and without
stickiness.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>