Add support for "LoadBalancer" Services
See issue #389.

Previously, when a Service resource is present in a manifest file, the
Kubernetes Controller would register it in the Krake Application, with
the IP taken from the kubeconfig file of the Cluster, and the external
port where it is reachable (called "node_port"). If no external port is
present, no service is registered. This works for the Kubernetes
Services of type "NodePort" and "ClusterIP".

However, in the case of a Service of type "LoadBalancer", a different
external IP is given by the Kubernetes cluster to the service. In this
case, the external port where the service is reachable also changes, and
is the "port" given in the resource specifications.

The "LoadBalancer" type is now supported, and when an IP has been given,
this IP and the port selected by the user are registered in the Krake
Application, to be read by the user.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>