Change scheduled timestamp semantic
Previously, the `scheduled` timestamp represented the time when the
last update of the application was scheduled. This was not very clear,
as this timestamp was used mostly for limiting the handling of an
Application by the Scheduler or the Kubernetes controller.

Thus, this timestamp was renamed to `kube_controller_triggered`, to
represent the role of the timestamp in the flow of an Application
through the Krake components.

At the same time, a new timestamp was added, which represent the last
time the Application was scheduled (but not deployed) to its current
scheduled cluster. This timestamp was named `scheduled`, to better
reflect what we could expect from a timestamp named like this.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>