Allow OPTIONS requests even without authentication
Previously, if an anonymous OPTIONS request for CORS was sent to the
API, it could be rejected because of the authentication method of Krake.
However, the CORS requests are usually sent by a browser, and may not
have any authentication. So if Krake has an authentication mechanism
set, the CORS requests would be refused with a 401.

To prevent this, all OPTIONS requests, which can only be handled in
Krake by the CORS handler, are now authenticated automatically as sent
by an anonymous user. This allow the CORS handler to get the requests,
instead of them being rejected directly.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>