Use the kubeconfig default namespace by default
See issue #403

Previously, if a resource had no Kubernetes namespace defined in its
manifest file, the namespace "default" would be used. However, a default
namespace other than "default" can be set in a kubeconfig file, which
the Kubernetes controller ignored.

Now, the Kubernetes client and the Kubernetes observer use the default
namespace set in the kubeconfig file if no namespace is set in the
manifest file. If one is set there, it has a higher priority.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>