Add timezone information to all timestamps
Previously most timestamps did not contain a timezone.
Some did however, and comparing timestamps with and without
timezones is not possible.

In the course of implementing new features in
select_kubernetes_cluster() in the scheduler, we
ended up having to compare two such timestamps, prompting
us to have to either remove the timezone from all timestamps
or add it to all. We opted for the latter, since using krake
in different timezones otherwise would cause problems.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>