Remove frequent rescheduling by fluctuating metric
When the scheduler has scheduled an application, this will
cause the kubernetes controller to update the application,
which will cause the scheduler to reconsider where to
schedule the application.

Previously, the scheduler would conclude to reschedule the
application, if the metrics had changed so that it would be
benefitial - independently of how recent the application had
been scheduled.

In the new implementation, the scheduler returns the current
cluster (assuming it is still valid based on the cluster
label constraints of the application), if it was less than
Scheduler.reschedule_after seconds ago that the application
was scheduled last.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>