Add timeouts to individual e2e test collections
Currently the task collecting the log files at the end
of a pipeline is not run if a task never returns, since
the pipeline runs into the CI gitlab project timeout.

We add a timeout to the pytest command for each
collection of e2e tests. This will cause the test to fail,
the log files to be collected and the pipeline to exit -
rather than run into the CI timeout of the gitlab project.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>